requirements.txt
supabase
psycopg2-binary
python-dotenv
requests
google-genai
